+++ src/test/scala/com/snowflake/snowpark_extensions/implicits/CaseExprExtensionsTest.scala
@@ -3,10 +3,10 @@ package com.snowflake.snowpark_extensions.implicits
 //Testing packages
 import com.snowflake.snowpark_extensions.implicits.Snowpark._
 import com.snowflake.snowpark_extensions.testutils.Serializer.df2Seq
-import org.scalatest.{FlatSpec, Matchers, Ignore}
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.should
 
-class CaseExprExtensionsTest extends FlatSpec with Matchers {
-behavior of "CaseExprExtensions class"
+class CaseExprExtensionsTest extends AnyFlatSpec with should.Matchers {
 
   "otherwise" should "match spark otherwise" in {
     val casSnowpark = CaseExprExtensionsSnowpark.test_otherwise()