+++ src/test/scala/com/snowflake/snowpark_extensions/implicits/SessionExtensionsTest.scala
@@ -3,10 +3,10 @@ package com.snowflake.snowpark_extensions.implicits
 //Testing packages
 import com.snowflake.snowpark_extensions.implicits.Snowpark._
 import com.snowflake.snowpark_extensions.testutils.Serializer.df2Seq
-import org.scalatest.{FlatSpec, Matchers}
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.should
 
-class SessionExtensionsTest extends FlatSpec with Matchers {
-  behavior of "SessionExtensions class"
+class SessionExtensionsTest extends AnyFlatSpec with should.Matchers {
 
   // "time" should "match spark time" in {
   //   val str = SessionExtensionsSnowpark.test_time()