+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-Version [0.1.0]
+Version [0.0.18]
 --------------
 - Implement helper for extensions 
 - Adding function extension rtrim
@@ -16,3 +16,10 @@ Version [0.1.0]
 - Adding function extension repeat
 - Adding function extension translate
 - Adding function extension and next_day
+
+Version [0.0.19]
+--------------
+- Updated dependency to snowpark 1.9.0
+- Adding function extension for Column.between
+- Adding function extension for DataFrame.alias
+- Extending DataFrame.selectExpr to support some alias cases